# Speed up read coverage collection