This code decreases the runtime for `samplot plot` by ~20% based on my test data. 

After some profiling I noticed that the `add_coverage` was consuming the majority of the resources. 

*Test command:*
```
samplot plot -z 4 -n HG002 AS=9 AP=18 control-sample:HG003 control-sample:HG004 --start_ci 0,0 --end_ci 0,0 -t DEL -c 19 -s 12694867 -e 12698924 -o tmp/DEL_19_12694867_12698924.png -d 1 -b test/data/HG002_Illumina.bam test/data/HG003_Illumina.bam test/data/HG004_Illumina.bam
```

<img width="888" alt="image" src="https://github.com/ryanlayer/samplot/assets/27061883/ee6ccfd5-5f6b-444c-a2c8-7cbb8bfecbde">


Specifically this loop in the `add_coverage` function was consuming the majority of the resources.

https://github.com/ryanlayer/samplot/blob/70cad157e7764c59a6149538305dcccaea8e11b9/samplot/samplot.py#L258-L267

Therefore I replaced the loop with using numpy for generate a matrix to represent the coverage over a defined range. Then I could just increment the coverage over the mapped regions of each read instead of updating each position separately. 

**Benchmarks**
*before*
```
Benchmark #1: samplot plot  -z 4 -n 'HG002 AS=9 AP=18' control-sample:HG003 control-sample:HG004 --start_ci '0,0' --end_ci '0,0' -t DEL -c 19 -s 12694867 -e 12698924 -o tmp/DEL_19_12694867_12698924.png -d 1 -b test/data/HG002_Illumina.bam test/data/HG003_Illumina.bam test/data/HG004_Illumina.bam
  Time (mean ± σ):      2.526 s ±  0.016 s    [User: 3.923 s, System: 0.308 s]
  Range (min … max):    2.501 s …  2.550 s    10 runs
 ```

*after*
```
Benchmark #1: samplot plot  -z 4 -n 'HG002 AS=9 AP=18' control-sample:HG003 control-sample:HG004 --start_ci '0,0' --end_ci '0,0' -t DEL -c 19 -s 12694867 -e 12698924 -o tmp/DEL_19_12694867_12698924.png -d 1 -b test/data/HG002_Illumina.bam test/data/HG003_Illumina.bam test/data/HG004_Illumina.bam
  Time (mean ± σ):      1.939 s ±  0.045 s    [User: 3.319 s, System: 0.322 s]
  Range (min … max):    1.901 s …  2.037 s    10 runs
```
Speedup = (2.526 -  1.939) / 2.526 = 0.2323832146